+++ src/main/java/pokeraidbot/commands/EggHatchedCommand.java
@@ -4,13 +4,19 @@ import com.jagrosh.jdautilities.commandclient.CommandEvent;
 import com.jagrosh.jdautilities.commandclient.CommandListener;
 import net.dv8tion.jda.core.entities.User;
 import org.apache.commons.lang3.ArrayUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import pokeraidbot.BotService;
 import pokeraidbot.domain.config.LocaleService;
 import pokeraidbot.domain.errors.UserMessedUpException;
+import pokeraidbot.domain.gym.Gym;
 import pokeraidbot.domain.gym.GymRepository;
+import pokeraidbot.domain.pokemon.Pokemon;
+import pokeraidbot.domain.pokemon.PokemonRaidInfo;
+import pokeraidbot.domain.pokemon.PokemonRaidStrategyService;
 import pokeraidbot.domain.pokemon.PokemonRepository;
+import pokeraidbot.domain.raid.Raid;
 import pokeraidbot.domain.raid.RaidRepository;
 import pokeraidbot.infrastructure.jpa.config.Config;
 import pokeraidbot.infrastructure.jpa.config.ServerConfigRepository;
@@ -26,17 +32,20 @@ public class EggHatchedCommand extends ConfigAwareCommand {
     private final PokemonRepository pokemonRepository;
     private final LocaleService localeService;
     private final BotService botService;
+    private final PokemonRaidStrategyService raidStrategyService;
 
     public EggHatchedCommand(GymRepository gymRepository, RaidRepository raidRepository,
                              PokemonRepository pokemonRepository, LocaleService localeService,
                              ServerConfigRepository serverConfigRepository,
-                             CommandListener commandListener, BotService botService) {
+                             CommandListener commandListener, BotService botService,
+                             PokemonRaidStrategyService raidStrategyService) {
         super(serverConfigRepository, commandListener, localeService);
         this.pokemonRepository = pokemonRepository;
         this.localeService = localeService;
         this.botService = botService;
+        this.raidStrategyService = raidStrategyService;
         this.name = "hatch";
-        this.help = "Report a raid egg hatched, and into what.";//localeService.getMessageFor(LocaleService.CHANGE_RAID_HELP, LocaleService.DEFAULT);
+        this.help = localeService.getMessageFor(LocaleService.EGG_HATCH_HELP, LocaleService.DEFAULT);
         this.gymRepository = gymRepository;
         this.raidRepository = raidRepository;
     }
@@ -46,14 +55,34 @@ public class EggHatchedCommand extends ConfigAwareCommand {
         final User user = commandEvent.getAuthor();
         final String userName = user.getName();
         final String[] args = commandEvent.getArgs().split(" ");
-        if (args.length != 2) {
+        if (args.length < 2) {
             throw new UserMessedUpException(userName,
                     localeService.getMessageFor(LocaleService.BAD_SYNTAX, localeService.getLocaleForUser(user)));
         }
         String pokemonName = args[0].trim().toLowerCase();
+        final String[] gymArguments = ArrayUtils.removeAll(args, 0);
+        String gymName = StringUtils.join(gymArguments, " ");
+        final String region = config.getRegion();
+        final Gym gym = gymRepository.search(user, gymName, region);
+        final Raid raid = raidRepository.getActiveRaidOrFallbackToExRaid(gym, region, user);
 
+        final Pokemon pokemon = pokemonRepository.search(pokemonName, user);
+        final PokemonRaidInfo existingRaidInfo = raidStrategyService.getRaidInfo(raid.getPokemon());
+        final int newBossTier = raidStrategyService.getRaidInfo(pokemon).getBossTier();
+        if (!raid.getPokemon().isEgg()) {
+            throw new UserMessedUpException(user,
+                    localeService.getMessageFor(LocaleService.EGG_ALREADY_HATCHED,
+                            localeService.getLocaleForUser(user), raid.getPokemon().toString()));
+        }
+        if (pokemon.isEgg() || newBossTier != existingRaidInfo.getBossTier()) {
+            throw new UserMessedUpException(user, localeService.getMessageFor(LocaleService.EGG_WRONG_TIER,
+                    localeService.getLocaleForUser(user)));
+        }
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("Trying to hatch raid " + raid + " into " + pokemonName);
+        }
         AlterRaidCommand.changePokemon(this, gymRepository, localeService, pokemonRepository, raidRepository,
                 botService.getTrackingCommandListener(), commandEvent, config, user, userName, pokemonName,
-                ArrayUtils.removeAll(args, 0));
+                gymArguments);
     }
 }